Создан презентер фильма. Удалены лишние комментарии
