+++ src/presenter/filmList.js
@@ -1,18 +1,16 @@
 import FilmsListView from '../view/films-list.js';
 import NoFilmsView from '../view/no-films.js';
-import CardFilmView from '../view/card-film.js';
 import ShowMoreView from '../view/show-more.js';
 import FilmsView from '../view/films.js';
-import {renderElement, isEscEvent} from '../utils/dom.js';
+import {renderElement} from '../utils/dom.js';
 import {Title} from '../utils/films.js'; //getNumberFilms
-import FilmDetailsView from '../view/film-details.js';
+import FilmPresenter from './film.js';
 
 const FILM_COUNT_PER_STEP = 5;
-const site = document.body;
 
 class FilmsBoard {
-  constructor(filmContainer) {
-    this._filmContainer = filmContainer;
+  constructor(filmsContainer) {
+    this._filmsContainer = filmsContainer;
     this._renderedFilmCount = FILM_COUNT_PER_STEP;
     // Т.к. основной список фильмов используется для отрисовки ShowMore кнопки, то сделаем её видимой для всего класса
     this._mainFilmsListComponent = null;
@@ -20,89 +18,27 @@ class FilmsBoard {
     this._filmsComponent = new FilmsView();
     this._noFilmsComponent = new NoFilmsView();
     this._showMoreComponent = new ShowMoreView();
-    this._filmDetailsComponent = null;
 
-    this._handleFilmDetails = this._handleFilmDetails.bind(this);
     this._handleShowMoreClick = this._handleShowMoreClick.bind(this);
-    this._onEscKeydown = this._onEscKeydown.bind(this);
-    this._handlePosterClick = this._handlePosterClick.bind(this);
   }
 
   init(films) {
     this._films = films.slice();
 
     // Отрисовываем контейнер, в котором будут списки фильмов
-    renderElement(this._filmContainer, this._filmsComponent.getElement());
+    renderElement(this._filmsContainer, this._filmsComponent.getElement());
     // Переходим к отрисовке содержимого
     this._renderFilmsBoard();
-  }
-
-  _handleFilmDetailsClick(evt) {
-    if (evt.target.tagName === 'BUTTON') {
-      evt.target.classList.toggle('film-details__control-button--active');
-    }
-  }
-
-  _handleFilmDetails() {
-    document.removeEventListener('keydown', this._onEscKeydown);
-    site.classList.remove('hide-overflow');
-    site.removeChild(this._filmDetailsComponent.getElement());
-  }
-
-  _onEscKeydown(evt) {
-    if (isEscEvent(evt)) {
-      evt.preventDefault();
-      this._handleFilmDetails();
-    }
-  }
-
-  _renderFilmDetails(film) {
-    this._filmDetailsComponent = new FilmDetailsView(film);
-
-    site.appendChild(this._filmDetailsComponent.getElement());
-    site.classList.add('hide-overflow');
-    document.addEventListener('keydown', this._onEscKeydown);
-
-    this._filmDetailsComponent.setOnFilmDetailsClick(this._handleFilmDetailsClick);
-    this._filmDetailsComponent.setOnCloseButtonClick(this._handleFilmDetails);
-
-    // todo При открытии нового попапа прежний закрывается, например при клике на другую карточку при открытом попапе.
-    // !Одновременно может быть открыт только один попап.
-    // !Несохранённые изменения (неотправленный комментарий) пропадают.
-  }
-
-  _checkClass(item) {
-    return item === 'film-card__controls-item';
-  }
-
-  _handlePosterClick(evt) {
-    const buttonClasses = Array.from(evt.target.classList);
-
-    if (buttonClasses.some(this._checkClass)) {
-      evt.target.classList.toggle('film-card__controls-item--active');
-    }
-  }
-
-  _renderFilm(filmsListElement, film) {
-    // Отрисовываем карточку фильма (мини)
-    // Т.к. нам нужно много разных карточек, то мы не выносим new CardFilmView в конструктор, а создаём "на месте"
-    const filmComponent = new CardFilmView(film);
-    // Обработчики кликов по названию, картинке и комментариям
-    filmComponent.setOnPosterClick(() => this._renderFilmDetails(film));
-    filmComponent.setOnTitleClick(() => this._renderFilmDetails(film));
-    filmComponent.setOnCommentsClick(() => this._renderFilmDetails(film));
-    // Обработчик клика по "нравится, смотрел, буду смотреть"
-    filmComponent.setOnControlsClick(this._handlePosterClick);
-    // Когда вся картчка отрисована, вставляем её в разметку
-    // ?На подумать. Зачем вставлять каждую карточку в разметку, когда можно отрисовать сразу 5 (или меньше, если их меньше) и вставить "блок"?
-    // ?Можно использовать document.createDocumentFragment()
-    // ??или вообще возвращать template, который потом отрисовывать на лист
-    renderElement(filmsListElement, filmComponent.getElement());
   }
 
   _renderNoFilms() {
     // Отрисовываем заглушку для списка без фильмов
     renderElement(this._filmsComponent.getElement(), this._noFilmsComponent.getElement());
+  }
+
+  _renderFilm(container, film) {
+    const filmPresenter = new FilmPresenter();
+    filmPresenter.init(container, film);
   }
 
   _renderMainFilmsList() {
